update confing (NOT WORKING)
